+++ CHANGES.rst
@@ -5,6 +5,8 @@
 * Fix comment submission anti-spam validation.
 * Fix bypassing of locked galleries via direct links.
 * Fix typo in MQ picture check.
+* Send pictures using Flask and kill all hardcoded domain name
+  references.
 
 1.0.0 (2015-03-16)
 ------------------